refactor utils-config dependency
