+++ pyhodl/config.py
@@ -62,8 +62,6 @@ HISTORICAL_DATA_FOLDER = os.path.join(
 
 DATE_TIME_FORMAT = "%Y-%m-%d %H:%M:%S %z"
 DATE_TIME_KEY = "datetime"
-LONG_DEC_FORMAT = "{0:.5f}"
-SHORT_DEC_FORMAT = "{0:.3f}"
 VALUE_KEY = "val"
 INFINITY = float("inf")
 NAN = float("nan")
@@ -96,3 +94,14 @@ def get_coin_historical_data_file(currency):
     """
 
     return os.path.join(HISTORICAL_DATA_FOLDER, currency.lower() + ".json")
+
+
+def is_crypto(coin):
+    """
+    :param coin: str or Coin
+        Coin to check
+    :return: bool
+        True iff coin is among crypto supported
+    """
+
+    return coin not in FIAT_COINS
